#821 [REFACTOR] shared/component/loading 분류 완료
