--- src/components/Loading/index.js
@@ -1,2 +1,0 @@
-export { default as FetchLoading } from './FetchLoading/FetchLoading';
-export { default as FetchLoadingOverlay } from './FetchLoadingOverlay/FetchLoadingOverlay';
