+++ src/pages/ExamReviewsPage/components/SearchExamReviews/SearchExamReviewsSuspense.jsx
@@ -3,7 +3,8 @@ import { useSearchParams } from 'react-router-dom';
 import { ErrorBoundary } from 'react-error-boundary';
 import { QueryErrorResetBoundary } from '@tanstack/react-query';
 
-import { FetchLoading } from '@/components';
+import { FetchLoading } from '@/shared/component';
+
 import {
   SearchExamReviews,
   SearchExamReviewsErrorFallback,
