+++ src/pages/LoginPage/FindIdPage/FindIdPage.jsx
@@ -3,10 +3,9 @@ import { useNavigate } from 'react-router-dom';
 
 import { useFindId } from '@/apis';
 
-import { Input, Submit } from '@/shared/component';
+import { FetchLoadingOverlay, Input, Submit } from '@/shared/component';
 import { Icon } from '@/components/Icon';
 
-import { FetchLoadingOverlay } from '@/components/Loading';
 import { LOADING_MESSAGE } from '@/constants';
 
 import {
