+++ src/pages/MyPage/pages/PointLogsPage/PointLogsPage.jsx
@@ -3,8 +3,8 @@ import { ErrorBoundary } from 'react-error-boundary';
 import { QueryErrorResetBoundary } from '@tanstack/react-query';
 
 import { useAuth } from '@/hooks';
-import { BackAppBar } from '@/shared/component';
-import { FetchLoading, Icon } from '@/components';
+import { BackAppBar, FetchLoading } from '@/shared/component';
+import { Icon } from '@/components';
 import {
   PointLogs,
   PointLogsErrorFallback,
