+++ src/pages/MyPage/pages/PointLogsPage/components/PointLogs/PointLogs.jsx
@@ -1,6 +1,8 @@
 import { getPointLogs } from '@/apis';
 import { useSuspensePagination } from '@/hooks';
-import { FetchLoading } from '@/components';
+
+import { FetchLoading } from '@/shared/component';
+
 import { PointLog } from '@/pages/MyPage/pages/PointLogsPage';
 import { flatPaginationCache } from '@/utils';
 import { QUERY_KEY } from '@/constants';
