+++ src/pages/PostsPage/components/Posts/PostsSuspense.jsx
@@ -2,7 +2,7 @@ import { Suspense } from 'react';
 import { ErrorBoundary } from 'react-error-boundary';
 import { QueryErrorResetBoundary } from '@tanstack/react-query';
 
-import { FetchLoading } from '@/components';
+import { FetchLoading } from '@/shared/component';
 import { Posts, PostsErrorFallback } from '@/pages/PostsPage';
 
 export default function PostsSuspense({ saveScrollPosition }) {
