+++ src/shared/component/loading/FetchLoading/FetchLoading.module.css
